Cleaned up products controllers, states, and routes
