+++ browser/js/products/products.js
@@ -11,17 +11,6 @@ app.config(function ($stateProvider) {
         }
     });
 
-    $stateProvider.state('product', {
-        url: '/products/:productId',
-        controller: 'ProductController',
-        templateUrl: 'js/products/product.html',
-        resolve: {
-          theProduct: function (ProductsFactory, $stateParams) {
-            return ProductsFactory.fetchById($stateParams.productId);
-          }
-        }
-    });
-
 });
 
 app.controller('ProductsController', function ($scope, allProducts) {
@@ -29,10 +18,3 @@ app.controller('ProductsController', function ($scope, allProducts) {
     $scope.products = allProducts;
 
 });
-
-app.controller('ProductController', function ($scope, theProduct) {
-
-  $scope.product = theProduct;
-
-});
-
